added improvements according to AI comments
