+++ src/App.tsx
@@ -1,8 +1,32 @@
 import React from 'react';
-import { Component } from './Component';
 
-export const App: React.FC = () => (
-  <div className="App">
-    <Component />
-  </div>
-);
+type State = {
+  pressedKey: string | null;
+};
+export class App extends React.Component<State> {
+  state: State = {
+    pressedKey: '',
+  };
+
+  handleKeyUp = (event: KeyboardEvent) => {
+    this.setState({ pressedKey: event.key });
+  };
+
+  componentDidMount(): void {
+    document.addEventListener('keyup', this.handleKeyUp);
+  }
+
+  componentWillUnmount(): void {
+    document.removeEventListener('keyup', this.handleKeyUp);
+  }
+
+  render() {
+    return this.state.pressedKey ? (
+      <p className="App__message">
+        The last pressed key is [{this.state.pressedKey}]
+      </p>
+    ) : (
+      <p className="App__message">Nothing was pressed yet</p>
+    );
+  }
+}
