--- src/Component.tsx
@@ -1,33 +1,0 @@
-import React from 'react';
-
-type Props = {};
-type State = {
-  pressedKey: string;
-};
-export class Component extends React.Component<Props, State> {
-  state: State = {
-    pressedKey: '',
-  };
-
-  handleKeyUp = (event: KeyboardEvent) => {
-    this.setState({ pressedKey: event.key });
-  };
-
-  componentDidMount(): void {
-    document.addEventListener('keyup', this.handleKeyUp);
-  }
-
-  componentWillUnmount(): void {
-    document.removeEventListener('keyup', this.handleKeyUp);
-  }
-
-  render() {
-    return this.state.pressedKey ? (
-      <p className="App__message">
-        The last pressed key is [{this.state.pressedKey}]
-      </p>
-    ) : (
-      <p className="App__message">Nothing was pressed yet</p>
-    );
-  }
-}
