+++ tsconfig.json
@@ -1,8 +1,6 @@
 {
   "extends": "@mate-academy/students-ts-config",
-  "include": [
-    "src"
-  ],
+  "include": ["src"],
   "compilerOptions": {
     "sourceMap": false,
     "types": ["node", "cypress"]
